Split OnboardingScreen.kt into 6 different files for readability
- Deleted unnecessary code
